reorg non-fixture test helpers; use harry potter helpers instead of cats :(
