+++ tests/framework/results/test_manager.py
@@ -7,19 +7,21 @@ from loguru import logger
 from vivarium.framework.results.manager import ResultsManager
 from vivarium.interface.interactive import InteractiveContext
 
-from ...helper_components import (
-    CatActivityObserver,
-    CatResultsStratifier,
-    CatToyObserver,
-)
-from .mocks import (
+from .helpers import (
     BIN_BINNED_COLUMN,
     BIN_LABELS,
     BIN_SILLY_BINS,
     BIN_SOURCE,
     CATEGORIES,
+    CONFIG,
+    FAMILIARS,
     NAME,
+    POWER_LEVELS,
     SOURCES,
+    STUDENT_HOUSES,
+    FamiliarObserver,
+    HogwartsResultsStratifier,
+    StudentHouseObserver,
     mock_get_value,
     sorting_hat_serial,
     sorting_hat_vector,
@@ -263,33 +265,31 @@ def test_stratified_metrics_initialized_as_zeros_dataframes():
     """
 
     components = [
-        CatToyObserver(),
-        CatActivityObserver(),
-        CatResultsStratifier(),
+        StudentHouseObserver(),
+        FamiliarObserver(),
+        HogwartsResultsStratifier(),
     ]
-    config = {
-        "stratification": {
-            "default": ["personality", "favorite_toy"],
-        },
-    }
-    sim = InteractiveContext(configuration=config, components=components)
+
+    sim = InteractiveContext(configuration=CONFIG, components=components)
     metrics = sim._results.metrics
     assert isinstance(metrics, dict)
-    assert set(metrics) == set(["cat_toy", "cat_activity"])
+    assert set(metrics) == set(["student_house", "familiar"])
     for metric in metrics:
         result = metrics[metric]
         assert isinstance(result, pd.DataFrame)
         assert result.columns == ["value"]
         assert result["value"].unique() == [0.0]
-    assert metrics["cat_toy"].index.equals(
+    STUDENT_HOUSES_LIST = list(STUDENT_HOUSES)
+    POWER_LEVELS_STR = [str(lvl) for lvl in POWER_LEVELS]
+    assert metrics["student_house"].index.equals(
         pd.MultiIndex.from_product(
-            [["psycopath", "cantankerous"], ["string", "human_face"]],
-            names=["personality", "favorite_toy"],
+            [STUDENT_HOUSES_LIST, POWER_LEVELS_STR],
+            names=["student_house", "power_level"],
         )
     )
-    assert metrics["cat_activity"].index.equals(
+    assert metrics["familiar"].index.equals(
         pd.MultiIndex.from_product(
-            [["psycopath", "cantankerous"], ["sleep", "eat"]],
-            names=["personality", "favorite_activity"],
+            [FAMILIARS, POWER_LEVELS_STR],
+            names=["familiar", "power_level"],
         )
     )
